backend/requirements.txt
flask
flask-cors
neo4j
python-dotenv
requests
